Fix buffer overflow on ALSA MIDI output
This fixes a buffer overflow that occurs when stopping MIDI playback
under Windows 95 with PCem's ALSA RawMidi interface.

While it removes all the code that keeps track of the state of the MIDI
messages, it seems like ALSA keeps its own buffer for it [1].  In my own
testing I couldn't spot any differences between this approach and
keeping track of the state.

[1]: https://www.alsa-project.org/alsa-doc/alsa-lib/rawmidi.html